split up SecureHeadersTest into categories; add some test cases
