+++ tests/SafeModeTest.php
@@ -1,0 +1,106 @@
+<?php
+
+namespace Tests;
+
+use Aidantwoods\SecureHeaders\Http\StringHttpAdapter;
+use Aidantwoods\SecureHeaders\SecureHeaders;
+use PHPUnit_Framework_TestCase;
+
+class SafeModeTest extends PHPUnit_Framework_TestCase
+{
+    private $assertions = array(
+        'Contains',
+        'NotContains',
+        'Equals',
+        'Regexp',
+        'NotRegExp'
+    );
+
+    function dataSafeMode()
+    {
+        return array(
+            array(
+                'test' => 
+                    function(&$headers){
+                        $headers->hsts(31536000, true, true);
+                    },
+                'assertions' => array(
+                    'Contains' =>
+                        'Strict-Transport-Security: max-age=31536000; includeSubDomains; preload'
+                )
+            ),
+            array(
+                'test' => 
+                    function(&$headers){
+                        $headers->safeMode();
+                        $headers->hsts(31536000, true, true);
+                    },
+                'assertions' => array(
+                    'NotContains' =>
+                        'Strict-Transport-Security: max-age=31536000; includeSubDomains; preload',
+                    'Contains' =>
+                        'Strict-Transport-Security: max-age=86400'
+                )
+            ),
+            array(
+                'test' => 
+                    function(&$headers){
+                        $headers->safeMode();
+                        $headers->strictMode();
+                    },
+                'assertions' => array(
+                    'NotContains' =>
+                        'Strict-Transport-Security: max-age=31536000; includeSubDomains; preload',
+                    'Contains' =>
+                        'Strict-Transport-Security: max-age=86400'
+                )
+            ),
+            array(
+                'test' => 
+                    function(&$headers){
+                        $headers->safeMode();
+                        $headers->hpkp('abcd', 31536000, true);
+                    },
+                'assertions' => array(
+                    'NotContains' =>
+                        'max-age=31536000; pin-sha256="abcd"; includeSubDomains',
+                    'Contains' =>
+                        'Public-Key-Pins: max-age=10; pin-sha256="abcd"'
+                )
+            )
+        );
+    }
+
+    /**
+     * @dataProvider dataSafeMode
+     * @param $test
+     * @param $assertions
+     */
+    public function testSafeMode($test, $assertions)
+    {
+        $headers = new SecureHeaders($headerStrings = new StringHttpAdapter);
+        $headers->errorReporting(false);
+        $test($headers);
+        $headers->done();
+
+        $headersString = $headerStrings->getSentHeaders();
+
+        foreach ($this->assertions as $assertion)
+        {
+            if (isset($assertions[$assertion]))
+            {
+                if ( ! is_array($assertions[$assertion]))
+                {
+                    $assertions[$assertion] = array($assertions[$assertion]);
+                }
+                foreach ($assertions[$assertion] as $assertionString)
+                {
+                    $this->{'assert'.$assertion}(
+                        $assertionString,
+                        $headersString
+                    );
+                }
+            }
+        }
+    }
+}
